+++ plantparent/src/App.js
@@ -1,16 +1,26 @@
 import React from 'react';
 import './App.css';
-import Home from './Home';
-import logo from './logo.png';
+import Home from './components/Home';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import Navbar from './components/navbar';
+import Chat from './components/chat';
+import About from './components/about';
 
 
 function App() {
   return (
     
-    <div className="App">
-      <img src={logo} width={250} height={250} alt="Logo" />
-      <Home />
-    </div>
+    <Router>
+      <Navbar />
+      <div style={{ padding: '20px' }}>
+          <Routes>
+            <Route path ='/' element={<Home/>}/>
+            <Route path ='/Home' element={<Home/>}/>
+            <Route path ='/chat' element={<Chat/>}/>
+            <Route path ='/about' element={<About/>}/>
+          </Routes>
+      </div>
+    </Router>
   );
 }
 export default App;
