+++ plantparent/src/components/Home.css
