+++ plantparent/src/components/about.js
@@ -1,0 +1,8 @@
+function About() {
+    return (
+     <div>
+        <h2>Chat blank page</h2>
+     </div>
+    );
+  }
+  export default About;
