+++ plantparent/src/components/chat.js
@@ -1,0 +1,8 @@
+function Chat() {
+    return (
+     <div>
+        <h2>Chat blank page</h2>
+     </div>
+    );
+  }
+  export default Chat;
